requirements.txt
fastapi
uvicorn
python-dotenv
aiomysql
redis
pydantic
requests
langchain
langchain-core
langchain-groq